clear the useless comment
