+++ app/src/main/java/com/example/seamless/MainActivity.kt
@@ -46,14 +46,96 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    SeamlessApp()
-
+                    //SeamlessApp()
                 }
             }
         }
     }
 }
+//UI
 
+@Composable
+fun SettingPart()
+{
+    Row{
+        OutlinedButton(
+            //Colors = Color.Blue,
+            modifier = Modifier
+                .height(60.dp)
+                .padding(10.dp)
+                .size(100.dp)
+                ,
+            onClick = {
+            // enter the settings
+        })
+        {
+            Text("Setting")
+        }
+
+    }
+}
+@Composable
+fun FunctionList(functionList: List<Function>, modifier:Modifier = Modifier)
+{
+    LazyColumn (
+        modifier = modifier,
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    )
+    {
+        items(functionList){
+            function: Function -> FunctionCard(function = function,
+            modifier = Modifier.padding(20.dp),
+
+            )
+
+        }
+    }
+}
+
+@Composable
+fun MainPart()
+{
+    SettingPart()
+    FunctionList(
+        functionList = DataSource().loadFunction()
+    )
+}
+
+@Composable
+fun FunctionCard(function: Function, modifier: Modifier = Modifier)
+{
+    //Space()
+    Card(
+        modifier = modifier.padding(10.dp,30.dp)
+    )
+    {
+
+        Column (modifier = modifier)
+        {
+            Image(
+                painter = painterResource(id = function.imageResourceId),
+                contentDescription = stringResource(id = function.functionResourceId),
+                modifier = Modifier
+                    .fillMaxWidth().height(130.dp),
+                contentScale = ContentScale.Crop
+            )
+            Button(
+                modifier = Modifier.fillMaxSize(),
+                onClick = {
+                    // go to personal/business functions
+                }) {
+                Text(
+                    text = stringResource(id = function.functionResourceId),
+                    modifier = Modifier
+                        .height(40.dp),
+                    style = MaterialTheme.typography.headlineSmall
+                )
+            }
+
+        }
+    }
+}
 
 @Composable
 @Preview
